requirements.txt
tweepy==4.12.1
python-telegram-bot==20.0
python-dotenv==0.19.2